refactor: move appdata to config
